ddl: migrate test-infra to testify for ddl.testSerialDBSuite
